feat: [#105] Deduplicate findings for the same line
When one check finds multiple issues in the same line of the same type (e.g. shell quoting issues) at the end of the reporting checks are deduplicated.

This means only one finding of the same type and line will be displayed, even though it might occur multiple times on the same line. By doing so, the output stays concise and easier to work with while avoiding confusion about checks popping up multiple times for the same line. Also from implementation perspective this does not require any rego check to be adjusted for ensuring backward-compatibility.